+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     buildToolsVersion "21.1.2"
 
     defaultConfig {
-        applicationId "ello.co.ello"
+        applicationId "co.ello.ElloApp"
         minSdkVersion 16
         targetSdkVersion 21
         versionCode 4
