+++ app/src/main/java/co/ello/ElloApp/Reachability.java
@@ -1,0 +1,13 @@
+package co.ello.ElloApp;
+
+import android.content.Context;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
+
+public class Reachability {
+    public static boolean isNetworkConnected( Context context ) {
+        ConnectivityManager manager = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo info = manager.getActiveNetworkInfo();
+        return (info != null);
+    }
+}
